# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  delivery.orderRequest:
    properties:
      address:
        type: string
      comment:
        type: string
      personNum:
        type: integer
      phone:
        type: string
      price:
        type: number
      products:
        items:
          $ref: '#/definitions/models.OrderProduct'
        type: array
      userId:
        type: integer
    required:
    - address
    - personNum
    - phone
    - price
    - products
    - userId
    type: object
...
  models.Order:
    properties:
      address:
        type: string
      comment:
        type: string
      id:
        type: integer
      person_num:
        type: integer
      phone:
        type: string
      price:
        type: number
      products:
        items:
          $ref: '#/definitions/models.OrderProduct'
        type: array
      user_id:
        type: integer
    type: object
  models.OrderProduct:
    properties:
      count:
        type: integer
      id:
        type: integer
      orderId:
        type: integer
      productId:
        type: integer
    required:
    - count
    - productId
    type: object
...
  tools.Error:
    properties:
      error:
        type: string
    type: object
  tools.Message:
    properties:
      message:
        type: object
    type: object
host: localhost:5000
info:
  contact:
    name: API Support
  description: This is a SkyDelivery server for Technopark Project.
  license: {}
  title: Swagger SkyDelivery API
  version: "1.0"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: Creating Order
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tools.Error'
      security:
      - basicAuth: []
      summary: Create Order Route
      tags:
      - Order
  /orders/:order_id:
    delete:
      consumes:
      - application/json
      description: Creating Order
      parameters:
      - description: ID of order
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tools.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/tools.Error'
      security:
      - basicAuth: []
      summary: Create Order Route
      tags:
      - Order
    get:
      consumes:
      - application/json
      description: Creating Order
      parameters:
      - description: ID of order
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/tools.Error'
      security:
      - basicAuth: []
      summary: Create Order Route
      tags:
      - Order
  /orders/checkout:
    post:
      consumes:
      - application/json
      description: Creating Order
      parameters:
      - description: New order data
        in: body
        name: OrderReq
        required: true
        schema:
          $ref: '#/definitions/delivery.orderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tools.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/tools.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tools.Error'
      security:
      - basicAuth: []
      summary: Create Order Route
      tags:
      - Order
...
swagger: "2.0"